Clause_Simplification/requirements.txt
textstat
python-dotenv
colorama

groq>=0.4.0
streamlit>=1.30.0
plotly>=5.18.0

# LangChain and related extensions
langchain>=0.0.300
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu

# PDF processing
pymupdf>=1.23.0

# Document processing (optional)
python-docx>=1.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0